encapsulated apikey and audienceid
